Test refactored
