+++ src/cli/i18n-copy/_test_/testFileResult.js
